ovn-northd.c: Omit unused columns in SB_Global.
Connections and SSL columns are not used by ovn-northd. Omit them in
IDL.

This is not a functional problem, but it may hide incremental processing
problems, because when status changes in Connection table, which is
referenced by SB_Global, ovn-northd will receive notifications for
SB_Global change, which triggers recompute. Some tests may pass when
this happens, while it should have failed if the recompute were not
triggered. An example is the test case "testing propagate
Port_Binding.up to NB and OVS", which should have revealed the problem
that was fixed by commit 2fbad127902c ("northd.c: Fix Port_Binding.up
reset handling in I-P.") but it didn't.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Acked-by: Mark Michelson <mmichels@redhat.com>