northd.c: Fix Port_Binding.up reset handling in I-P.
With northd incremental processing, when Port_Binding.up is cleared
manually, e.g. in test case "testing propagate Port_Binding.up to NB and
OVS", the change is ignored, and the test case may fail (the reason why
the test passes sometimes is because it triggers recompute, and a
separate patch will avoid such recomputes).

This patch fixes it by resetting the value to "false" in the
en_sync_from_sb node.

Note: alternatively we may modify the test case, because this is not a
real problem in production, but the fix in code is simple without
obvious side-effect, so just kept the test case and handled it in the
code.

Fixes: b41d74fcb382 ("northd: Ignore "up" column change if ignore_lsp_down is true.")
Reported-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Han Zhou <hzhou@ovn.org>
Acked-by: Mark Michelson <mmichels@redhat.com>