ovn-northd.at: Fix the LSP incremental processing test case.
The test case intended to ensure there are no more than 3 failures in 10
runs. However, it used "break" to exit the loop whenever a failure is
encountered, end up with at most 1 failure and so the final check for
the failure count will always pass. It should use "continue" instead.
Also change the failure threshold from 3/10 to 5/10 to make the test
more stable after the fix.

Fixes: 8c30ba13869e ("ovn-northd.at: Fix unstable LSP incremental processing test.")
Signed-off-by: Han Zhou <hzhou@ovn.org>
Acked-by: Dumitru Ceara <dceara@redhat.com>