ovn-northd.at: Check and compare DB content for incremental processing.
Test cases of ovn-northd incremental processing focus more on the
effectiveness of incremental processing and less on the functional
correctness of features, because it would be redundant and hard to
maintain. However, it is also important to make sure the incremental
processing logic is correct. The existing feature tests may not be
sufficient for this purpose before there are so many events in those
tests that can trigger recompute, and we can't tell if the incremental
processing logic is covered by those test executions.

This patch provides a generic approach to check the correctness of
incremental processing by dumping related DB tables before and after
triggering a recompute, and compare the contents of the tables to make
sure no change is triggered by the recompute. If there is any change,
the incremental processing logic must be wrong. This way we can ensure
the correctness without repeating excessive checks already covered by
the feature tests.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Acked-by: Numan Siddique <numans@ovn.org>