tests: fixed "Mirror - remote" and "Mirror - local"
When we run ovn-nbctl --wait=hv, there is no guarantee that, after
ovn-nbctl returns, conf.db is already updated.

If that ovsdb-server process is very slow (or not scheduled on a busy
system), it might still take some time before the db gets properly
updated.

But even after the db gets updated, ovs-vswitchd has to get notified of
that db change and run mirror_configure to reconfigure the mirror.
And, again, it might take a while for ovs-vswitchd process to get
scheduled on a busy system.

To try to ensure that ovs-vswitchd had run at least one loop we use
an appctl command that connects to ovs-vswitchd.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>