ovn-controller: avoid monitoring wrong chassis
To register a chassis to ovn-sb, ovn-controller create a "name-uuid"
for the chassis, which it will send to ovn-sb. In return, ovn-sb creates
the uuid based on that "name".
Using this "name-uuid" as a uuid in a monitor request to ovn-sb results in
monitoring a non-existing chassis.
The system returns on track as soon as ovn-controller receives the chassis
update from ovn-sb, as it will then send a new monitor request to ovn-sb,
with the proper chassis uuid.
However, if some ports were already in ovn-sb when it receives the monitor request
with the wrong uuid, ovn-sb will notify ovn-controller to delete all those ports.
This cause the following sequence in ovn-controller
- notification of a new port
- deletion of that port
- notification of a new port

We now prevent sending wrong uuid to ovn-sb.

This issue was causing test "testing load-balancer template IPv6" to fail
in a flaky way, due to the following warning:
if_status|WARN|Trying to release unknown interface vm2

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>