ovn-util: Factor out struct sorted_addresses into sorted_array.
It's actually a generic wrapper and will be useful for upcoming commits.
This commit doesn't perform any functional changes but cleans up a bit
the implementation unifying the sorted_array cleanup.  Before the change
the caller had to track whether it should free the internal 'arr' field
or not.  It's now taken care of inside the sorted_array_destroy() API
which the user always has to call.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Acked-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Han Zhou <hzhou@ovn.org>
Acked-by: Ales Musil <amusil@redhat.com>