Expose distributed gateway port information in NB DB
In order for the CMS to know which Chassis a distributed gateway port
is bond to, this patch updates the ovn-northd daemon to populate the
Logical_Router_Port table with that information.

A new column called "status" has been added to the Logical_Router_Port
table and a key called "hosting-chassis" will be populated by ovn-northd
with the name of the Chassis that is currently hosting the distributed
port.

The new "status" column uses key-value style so it's more future proof
and can be easily extended.

Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2107515
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Mark Michelson <mmichels@redhat.com>