Fix missing flows in ls_in_dhcp_options table
When ha-chassis-group is updated, causing a new hv to become master,
port move to a different chassis, and flows get created in table ls_in_dhcp_options.
If pb->chassis is reported empty by sb (because set to [] by the hv being the
previous owner of the port) those flows were deleted.
However, when pb->chassis is finally updated by sb (to the new hv), flows were
not re-installed, as runtime_data was not seen as updated.

This caused some random failures in test "external logical port".

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>