system-tests: Make sure that IPv6 address is available right away
Instead of waiting for the address duplication check to finish
request that the check is not performed at all. With that
we don't have to wait for the "tentative" state to go away
and the address is usable right away. Remove check for the
link-local address where it is not needed as the wait was
taking ~2 second. As a bonus it speeds up the system-tests
execution by ~2 minutes.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
(cherry picked from commit b7d5dd275c6d5244b694a1d4bc571e0b7afc2006)