ci: Update apt cache before installing gcc-multilib.
Most of the packages are pre-installed in the container image we're
using.  The only exception is gcc-multilib.  Given that the container
images are pre-built (weekly), there's a window of up to a week in which
CI jobs might not be able to install gcc-multilib, e.g.:
https://github.com/dceara/ovn/actions/runs/6428804529/job/17456662002#step:8:129

To avoid that, update cache before installing gcc-multilib.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Acked-by: Ales Musil <amusil@redhat.com>