northd: Reset ls_datapath_group if not all chassis support it.
The correct upgrade procedure mentions that databases are upgraded
at the same time as ovn-northd, so there's nothing that doesn't allow
northd to write to new columns.  The feature flag reports instead what
features chassis are aware of.  In this specific case, if a chassis
doesn't know how to interpret the 'ls_datapaths' column, ovn-northd
should set it to NULL and use the deprecated 'datapath' one.

Fixes: ab7a8fe4c018 ("northd: introduce ls_datapath_group column in lb sb db table")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Acked-by: Ales Musil <amusil@redhat.com>